Clean up landing page code
