+++ pcweb/pages/index/components/landing_page_style.py
